pyproject.toml
[project]
name = "api-thesammy2010-com"
version = "0.1.0"
description = "Repo to handle API requests"
authors = [
    {name = "TheSammy2010",email = "43960199+thesammy2010@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg[binary,pool] (>=3.2.9,<4.0.0)",
    "ariadne (>=0.26.2,<0.27.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "cryptography (>=45.0.3,<46.0.0)",
    "google-auth (>=2.40.2,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "api"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
pytest = "^8.3.5"
pre-commit = "^4.2.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
